refactor and simplify app dynamic components
